requirements/common.txt
alembic==1.6.5
aniso8601==9.0.1
APScheduler==3.7.0
blinker==1.4
Bootstrap-Flask==1.6.0
certifi==2021.5.30
click==8.0.1
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.13.1
email-validator==1.1.2
Flask==2.0.1
Flask-Admin==1.5.8
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.17
tzlocal==2.1
Unidecode==1.2.0
urllib3==1.26.5
visitor==0.1.3
Werkzeug==2.0.1
WTForms==2.3.3
